Move Grammar to core
